Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int num1,num2;
+    float num1,num2;
     if(num1>num2){
         printf("True");    
     }
